bugfix: Fix extra space in filtered menu by using Dynamic to completely remove Column elements that shouldn't be shown
The issue here is that there doesn't seem to be any box form A such that:

	Column[{A, B, C}]

Displays with A taking up zero vertical height, i.e. as if it wasn't there. We
want such a thing to be able to dynamically hide styles from the cell insertion menu
that don't match the typed filter text.